requirements.txt
# Video Clips Extractor - Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# FastAPI & Web Server
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
aiofiles>=23.2.1

# =============================================================================
# Transcription
# =============================================================================
openai-whisper>=20231117
torch>=2.0.0

# =============================================================================
# LLM Providers
# =============================================================================
ollama>=0.1.0
groq>=0.4.0
together>=1.0.0
httpx>=0.25.0

# =============================================================================
# Video Processing
# =============================================================================
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# =============================================================================
# YouTube Download
# =============================================================================
yt-dlp>=2024.1.0

# =============================================================================
# Google Drive Upload
# =============================================================================
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# =============================================================================
# Utilities
# =============================================================================
tqdm>=4.65.0
python-dotenv>=1.0.0

# =============================================================================
# Note: You also need FFmpeg installed on your system
# Windows: choco install ffmpeg OR download from https://ffmpeg.org/download.html
# Mac: brew install ffmpeg
# Linux: sudo apt install ffmpeg
# =============================================================================
